Move token entities to balances route (#111)
- Currently, token-info and token-type are only being used in the balances route – they might end up being used later by other routes but at that point we can move them to a "common" entity layer which can be shared between routes
- Removed tokenInfoSchema since we were not using it – currently we only validate the JSON coming from external services
