+++ src/routes/balances/openapi/api-balance.ts
@@ -1,6 +1,6 @@
 import { Balance as DomainBalance } from '../entities/balance.entity';
 import { ApiProperty } from '@nestjs/swagger';
-import { TokenInfo as ApiTokenInfo } from '../../../common/openapi/api-token-info';
+import { TokenInfo as ApiTokenInfo } from './api-token-info';
 
 export class Balance implements DomainBalance {
   @ApiProperty()
